ADD separation between GUI and logic. ADD new javaDoc
